I've made more edits trying to figure out why I can't get the contents of the list to output.

+++ Shoes.cs
@@ -9,7 +9,7 @@ namespace KnowledgeCheck2
 {
     public class Shoes
     {
-        public decimal Size;
+        public double Size;
         //public string Type;
     }
 }
